requirements-txt.txt
langchain-community==0.0.16
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.12.0
spacy==3.7.2
langchain==0.1.0
langchain-openai==0.0.5
streamlit==1.31.0
python-dotenv==1.0.0
pypdf==3.17.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
